Implement save versioning and minimal save data
